refactor: refactor mcp gateway to core and nacos impl
