+++ auto-configurations/spring-ai-alibaba-autoconfigure-mcp-gateway-nacos/src/main/resources/META-INF/spring-configuration-metadata.json
