+++ spring-ai-alibaba-bom/pom.xml
@@ -96,6 +96,13 @@
                 <artifactId>spring-ai-alibaba-core</artifactId>
                 <version>${project.version}</version>
             </dependency>
+            
+            <!-- Spring AI Alibaba Mcp Gateway Core -->
+            <dependency>
+                <groupId>com.alibaba.cloud.ai</groupId>
+                <artifactId>spring-ai-alibaba-mcp-gateway-core</artifactId>
+                <version>${project.version}</version>
+            </dependency>
 
             <!-- Spring AI Alibaba Starter -->
             <dependency>
@@ -142,7 +149,7 @@
 
             <dependency>
                 <groupId>com.alibaba.cloud.ai</groupId>
-                <artifactId>spring-ai-alibaba-starter-nacos-mcp-gateway</artifactId>
+                <artifactId>spring-ai-alibaba-starter-mcp-gateway-nacos</artifactId>
                 <version>${project.version}</version>
             </dependency>
